[iris] consolidate reconcile bench into benchmark_controller, drop e2e bench
Fold benchmark_reconcile.py into scripts/benchmark_controller.py as a
'reconcile' subcommand on its existing click group. Delete the unreferenced
tests/e2e/benchmark_controller.py, which was superseded by the
checkpoint-driven scripts/benchmark_controller.py.
